toFutureTimeTableDto() 메소드 이동
