+++ src/main/java/umc/forgrad/converter/StudentConverter.java
@@ -1,11 +1,7 @@
 package umc.forgrad.converter;
 
 import umc.forgrad.domain.Student;
-import umc.forgrad.domain.Subject;
 import umc.forgrad.dto.student.StudentResponseDto;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class StudentConverter {
 
@@ -19,21 +15,6 @@ public class StudentConverter {
         return StudentResponseDto.LoginResponseDto.builder()
                 .message(loginResultMessage)
                 .build();
-    }
-
-    public static List<StudentResponseDto.FutureTimeTableDto> toFutureTimeTableDto(List<Subject> subjectList) {
-
-        List<StudentResponseDto.FutureTimeTableDto> futureTimeTableDtoList = new ArrayList<>();
-
-        subjectList.forEach(subject -> futureTimeTableDtoList.add(StudentResponseDto.FutureTimeTableDto.builder()
-                .majorType(subject.getType())
-                .subject(subject.getName())
-                .grades(subject.getCredit())
-                .build()
-        ));
-
-        return futureTimeTableDtoList;
-
     }
 
 }
